add moved file
