+++ cmd/swagger/commands/version_number.go
@@ -1,5 +1,5 @@
 package commands
 
 func init() {
-	Version = "0.5.0-38-g9a0380c"
+	Version = "0.5.0-38-g5972946"
 }
