--- examples/tutorials/todo-list/server-complete/cmd/todo-list-server/doc.go
@@ -1,22 +1,0 @@
-/*Package main A To Do list application
-
-The product of a tutorial on goswagger.io
-
-
-    Schemes:
-      http
-    Host: localhost
-    BasePath: /
-    Version: 1.0.0
-
-    Consumes:
-    - application/io.goswagger.examples.todo-list.v1+json
-
-
-    Produces:
-    - application/io.goswagger.examples.todo-list.v1+json
-
-
-swagger:meta
-*/
-package main
